chore: Move AnchorService types
