+++ packages/core/src/anchor/ethereum/ethereum-anchor-service.ts
@@ -1,14 +1,10 @@
 import { CID } from 'multiformats/cid'
 import {
   CeramicApi,
-  AnchorService,
-  AnchorServiceAuth,
-  AuthenticatedAnchorService,
   DiagnosticsLogger,
   fetchJson,
   FetchRequest,
   AnchorEvent,
-  AnchorValidator,
 } from '@ceramicnetwork/common'
 import { StreamID } from '@ceramicnetwork/streamid'
 import {
@@ -29,6 +25,7 @@ import { AnchorRequestCarFileReader } from '../anchor-request-car-file-reader.js
 import { CASResponseOrError, ErrorResponse, AnchorRequestStatusName } from '@ceramicnetwork/codecs'
 import { decode } from 'codeco'
 import { EthereumAnchorValidator } from './ethereum-anchor-validator.js'
+import { AnchorService, AnchorServiceAuth, AnchorValidator, AuthenticatedAnchorService } from '../anchor-service.js'
 
 const DEFAULT_POLL_INTERVAL = 60_000 // 60 seconds
 const MAX_POLL_TIME = 86_400_000 // 24 hours
@@ -278,6 +275,10 @@ export class EthereumAnchorService implements AnchorService {
 
   private _updateEvents(): MonoTypeOperatorFunction<AnchorEvent> {
     return tap((event) => this.#events.next(event))
+  }
+
+  async close(): Promise<void> {
+    // Do Nothing
   }
 }
